+++ src/auth/jwt_manager.py
@@ -1,86 +1,64 @@
 """
-JWT Token Manager for Osservatorio ISTAT Data Platform
+Simplified JWT Token Manager for MVP - Issue #153
 
-Provides secure JWT token generation and validation:
-- RS256/HS256 JWT token support
-- Configurable token expiration
-- Refresh token management with SQLite storage
-- Token blacklisting for secure logout
-- Scope-based claims validation
+Basic JWT token generation and validation for MVP:
+- HS256 JWT token support only (simplified)
+- Basic token expiration
+- No refresh tokens (removed for MVP)
+- No token blacklisting (removed for MVP)
+- Simple claims validation
 """
 
-import hashlib
 import secrets
 from datetime import datetime, timedelta
 from typing import Optional
 
 try:
     import jwt
-    from cryptography.hazmat.primitives import serialization
-    from cryptography.hazmat.primitives.asymmetric import rsa
 except ImportError:
-    raise ImportError(
-        "JWT dependencies not installed. Run: pip install PyJWT cryptography"
-    )
+    raise ImportError("JWT dependencies not installed. Run: pip install PyJWT")
 
-from src.database.sqlite.manager_factory import get_audit_manager
 from src.utils.config import get_config
 from src.utils.logger import get_logger
 
 from .models import APIKey, AuthToken, TokenClaims
 
 logger = get_logger(__name__)
-# Security manager imported above
 
 
 class JWTManager:
-    """JWT token manager with SQLite backend for refresh tokens"""
+    """Simplified JWT token manager for MVP - Issue #153"""
 
     # Default token configuration
     DEFAULT_ACCESS_TOKEN_EXPIRE_MINUTES = 60  # 1 hour
-    DEFAULT_REFRESH_TOKEN_EXPIRE_DAYS = 30  # 30 days
 
-    # JWT algorithms supported
-    ALGORITHM_HS256 = "HS256"  # Symmetric
-    ALGORITHM_RS256 = "RS256"  # Asymmetric
+    # JWT algorithm (MVP: only HS256)
+    ALGORITHM_HS256 = "HS256"
 
     def __init__(
         self,
         db_path: Optional[str] = None,
         secret_key: Optional[str] = None,
-        algorithm: str = ALGORITHM_HS256,
     ):
-        """Initialize JWT manager
+        """Initialize simplified JWT manager for MVP
 
         Args:
-            db_path: SQLite database path for refresh token storage
+            db_path: Optional database path (kept for compatibility)
             secret_key: JWT signing secret (if None, will generate or load from config)
-            algorithm: JWT algorithm (HS256 or RS256)
         """
-        # Initialize specialized manager for audit and token operations
-        self.audit_manager = get_audit_manager(db_path)
-        self.algorithm = algorithm
+        self.algorithm = self.ALGORITHM_HS256
         self.logger = logger
 
-        # Initialize signing configuration
-        if algorithm == self.ALGORITHM_RS256:
-            self._init_rsa_keys()
-        else:
-            self._init_secret_key(secret_key)
+        # Initialize secret key only (no RSA for MVP)
+        self._init_secret_key(secret_key)
 
         # Load configuration
         config = get_config()
         self.access_token_expire_minutes = config.get(
             "jwt_access_token_expire_minutes", self.DEFAULT_ACCESS_TOKEN_EXPIRE_MINUTES
         )
-        self.refresh_token_expire_days = config.get(
-            "jwt_refresh_token_expire_days", self.DEFAULT_REFRESH_TOKEN_EXPIRE_DAYS
-        )
 
-        # Ensure JWT schema exists
-        self._ensure_jwt_schema()
-
-        logger.info(f"JWT Manager initialized with {algorithm} algorithm")
+        logger.info("Simplified JWT Manager initialized for MVP - Issue #153")
 
     def _init_secret_key(self, secret_key: Optional[str]):
         """Initialize HMAC secret key for HS256"""
@@ -98,151 +76,30 @@ class JWTManager:
                     "Generated new JWT secret key. Set JWT_SECRET_KEY environment variable for production."
                 )
 
-    def _init_rsa_keys(self):
-        """Initialize RSA key pair for RS256"""
-        # For production, these should be loaded from secure storage
-        # Here we generate them for demonstration
-        private_key = rsa.generate_private_key(public_exponent=65537, key_size=2048)
-
-        self.private_key = private_key.private_bytes(
-            encoding=serialization.Encoding.PEM,
-            format=serialization.PrivateFormat.PKCS8,
-            encryption_algorithm=serialization.NoEncryption(),
-        )
-
-        self.public_key = private_key.public_key().public_bytes(
-            encoding=serialization.Encoding.PEM,
-            format=serialization.PublicFormat.SubjectPublicKeyInfo,
-        )
-
-        logger.warning(
-            "Generated RSA keys in memory. Use persistent keys for production."
-        )
-
-    def _ensure_jwt_schema(self):
-        """Ensure JWT schema exists in SQLite database"""
-        try:
-            with self.audit_manager.transaction() as conn:
-                cursor = conn.cursor()
-
-                # Ensure api_credentials table exists first
-                cursor.execute(
-                    """
-                    CREATE TABLE IF NOT EXISTS api_credentials (
-                        id INTEGER PRIMARY KEY AUTOINCREMENT,
-                        service_name TEXT NOT NULL UNIQUE,
-                        api_key_hash TEXT NOT NULL,
-                        api_secret_hash TEXT,
-                        endpoint_url TEXT,
-                        is_active BOOLEAN DEFAULT 1,
-                        expires_at TIMESTAMP,
-                        last_used TIMESTAMP,
-                        usage_count INTEGER DEFAULT 0,
-                        rate_limit INTEGER DEFAULT 100,
-                        created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                        updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
-                    )
-                """
-                )
-
-                # Create refresh_tokens table
-                cursor.execute(
-                    """
-                    CREATE TABLE IF NOT EXISTS refresh_tokens (
-                        id INTEGER PRIMARY KEY AUTOINCREMENT,
-                        api_key_id INTEGER NOT NULL,
-                        token_hash TEXT NOT NULL UNIQUE,
-                        expires_at TIMESTAMP NOT NULL,
-                        is_revoked INTEGER DEFAULT 0,
-                        created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                        revoked_at TIMESTAMP NULL,
-                        FOREIGN KEY (api_key_id) REFERENCES api_credentials (id)
-                    )
-                """
-                )
-
-                # Create token_blacklist table for revoked JWT tokens
-                cursor.execute(
-                    """
-                    CREATE TABLE IF NOT EXISTS token_blacklist (
-                        id INTEGER PRIMARY KEY AUTOINCREMENT,
-                        token_hash TEXT NOT NULL UNIQUE,
-                        expires_at TIMESTAMP NOT NULL,
-                        reason TEXT,
-                        created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
-                    )
-                """
-                )
-
-                # Create indexes for performance
-                cursor.execute(
-                    """
-                    CREATE INDEX IF NOT EXISTS idx_refresh_tokens_api_key
-                    ON refresh_tokens(api_key_id)
-                """
-                )
-
-                cursor.execute(
-                    """
-                    CREATE INDEX IF NOT EXISTS idx_refresh_tokens_hash
-                    ON refresh_tokens(token_hash)
-                """
-                )
-
-                cursor.execute(
-                    """
-                    CREATE INDEX IF NOT EXISTS idx_refresh_tokens_expires
-                    ON refresh_tokens(expires_at)
-                """
-                )
-
-                cursor.execute(
-                    """
-                    CREATE INDEX IF NOT EXISTS idx_blacklist_hash
-                    ON token_blacklist(token_hash)
-                """
-                )
-
-                cursor.execute(
-                    """
-                    CREATE INDEX IF NOT EXISTS idx_blacklist_expires
-                    ON token_blacklist(expires_at)
-                """
-                )
-
-                conn.commit()
-                logger.debug("JWT schema ensured successfully")
-
-        except Exception as e:
-            logger.error(f"Failed to ensure JWT schema: {e}")
-            raise
-
     def create_access_token(
         self, api_key: APIKey, custom_claims: Optional[dict] = None
     ) -> AuthToken:
-        """Create JWT access token for API key
+        """Create simplified JWT access token for API key
 
         Args:
             api_key: APIKey object to create token for
-            custom_claims: Additional claims to include
+            custom_claims: Additional claims to include (optional)
 
         Returns:
-            AuthToken with access and refresh tokens
+            AuthToken with access token only (no refresh token for MVP)
         """
         try:
             now = datetime.utcnow()
             expires_at = now + timedelta(minutes=self.access_token_expire_minutes)
 
-            # Build token claims
+            # Build simplified token claims
             claims = {
                 "sub": str(api_key.id),  # Subject (API key ID)
                 "iss": "osservatorio-istat",  # Issuer
                 "aud": "osservatorio-api",  # Audience
                 "exp": expires_at,  # Expiration
                 "iat": now,  # Issued at
-                "scope": " ".join(api_key.scopes),  # Scopes as space-separated string
                 "api_key_name": api_key.name,
-                "rate_limit": api_key.rate_limit,
                 "jti": secrets.token_urlsafe(16),  # JWT ID for tracking
             }
 
@@ -250,308 +107,176 @@ class JWTManager:
             if custom_claims:
                 claims.update(custom_claims)
 
-            # Sign token
-            signing_key = (
-                self.private_key
-                if self.algorithm == self.ALGORITHM_RS256
-                else self.secret_key
-            )
-            access_token = jwt.encode(claims, signing_key, algorithm=self.algorithm)
+            # Create JWT token
+            access_token = jwt.encode(claims, self.secret_key, algorithm=self.algorithm)
 
-            # Create refresh token
-            refresh_token = self._create_refresh_token(api_key.id)
-
-            # Calculate expires_in for client
-            expires_in = int((expires_at - now).total_seconds())
-
-            return AuthToken(
+            # Create AuthToken object (no refresh token for MVP)
+            token = AuthToken(
                 access_token=access_token,
                 token_type="bearer",
-                expires_in=expires_in,
-                refresh_token=refresh_token,
-                scope=" ".join(api_key.scopes),
+                expires_in=self.access_token_expire_minutes * 60,
+                scope=" ".join(api_key.scopes) if api_key.scopes else "read",
             )
+
+            logger.debug(f"Access token created for API key: {api_key.name}")
+            return token
 
         except Exception as e:
             logger.error(f"Failed to create access token: {e}")
-            raise RuntimeError(f"Token creation failed: {e}")
+            raise
 
     def verify_token(self, token: str) -> Optional[TokenClaims]:
-        """Verify JWT token and extract claims
+        """Verify and decode JWT token
 
         Args:
-            token: JWT token to verify
+            token: JWT token string
 
         Returns:
-            TokenClaims if valid, None if invalid/expired
+            TokenClaims if token is valid, None if invalid/expired
         """
         try:
-            # Check if token is blacklisted
-            if self._is_token_blacklisted(token):
-                logger.warning("Attempted use of blacklisted token")
-                return None
-
-            # Verify token signature and expiration
-            signing_key = (
-                self.public_key
-                if self.algorithm == self.ALGORITHM_RS256
-                else self.secret_key
-            )
+            # Decode and verify token (MVP: disable audience validation for simplicity)
             payload = jwt.decode(
                 token,
-                signing_key,
+                self.secret_key,
                 algorithms=[self.algorithm],
-                audience="osservatorio-api",
-                issuer="osservatorio-istat",
+                options={"verify_aud": False},  # Disable audience verification for MVP
             )
 
-            # Extract claims
-            return TokenClaims(
+            # Extract claims (fixed field names for TokenClaims model)
+            token_claims = TokenClaims(
                 sub=payload.get("sub"),
-                iss=payload.get("iss"),
-                aud=payload.get("aud"),
-                exp=datetime.fromtimestamp(payload.get("exp", 0)),
-                iat=datetime.fromtimestamp(payload.get("iat", 0)),
+                iss=payload.get("iss", "osservatorio-istat"),
+                aud=payload.get("aud", "osservatorio-api"),
+                exp=datetime.fromtimestamp(payload.get("exp", 0))
+                if payload.get("exp")
+                else None,
+                iat=datetime.fromtimestamp(payload.get("iat", 0))
+                if payload.get("iat")
+                else None,
                 scope=payload.get("scope", "read"),
                 api_key_name=payload.get("api_key_name"),
-                rate_limit=payload.get("rate_limit", 100),
             )
 
+            logger.debug(
+                f"Token verified successfully for: {token_claims.api_key_name}"
+            )
+            return token_claims
+
         except jwt.ExpiredSignatureError:
-            logger.debug("Token expired")
+            logger.warning("JWT token has expired")
             return None
         except jwt.InvalidTokenError as e:
-            logger.debug(f"Invalid token: {e}")
+            logger.warning(f"Invalid JWT token: {e}")
             return None
         except Exception as e:
             logger.error(f"Token verification failed: {e}")
             return None
 
-    def refresh_access_token(self, refresh_token: str) -> Optional[AuthToken]:
-        """Create new access token using refresh token
+    def create_token_for_user(
+        self, user_id: str, username: str, scopes: Optional[list] = None
+    ) -> str:
+        """Create simple JWT token for basic user authentication (MVP helper)
 
         Args:
-            refresh_token: Refresh token to use
+            user_id: User identifier
+            username: Username
+            scopes: Optional list of scopes
 
         Returns:
-            New AuthToken if refresh successful, None otherwise
+            JWT token string
         """
         try:
-            # Verify refresh token
-            api_key_id = self._verify_refresh_token(refresh_token)
-            if not api_key_id:
-                return None
+            now = datetime.utcnow()
+            expires_at = now + timedelta(minutes=self.access_token_expire_minutes)
 
-            # Get API key details
-            # Note: Would need to add method to get API key by ID
-            # For now, create minimal APIKey object
-            api_key = APIKey(id=api_key_id, name="refreshed", scopes=["read"])
+            claims = {
+                "sub": user_id,
+                "username": username,
+                "exp": expires_at,
+                "iat": now,
+                "iss": "osservatorio-istat",
+                "scopes": scopes or ["read"],
+                "jti": secrets.token_urlsafe(16),
+            }
 
-            # Create new access token
-            return self.create_access_token(api_key)
-
-        except Exception as e:
-            logger.error(f"Token refresh failed: {e}")
-            return None
-
-    def revoke_token(self, token: str, reason: str = "manual_revocation") -> bool:
-        """Revoke (blacklist) a JWT token
-
-        Args:
-            token: JWT token to revoke
-            reason: Reason for revocation
-
-        Returns:
-            True if successfully revoked, False otherwise
-        """
-        try:
-            # Extract token ID (jti) for blacklisting
-            payload = jwt.decode(token, options={"verify_signature": False})
-            jti = payload.get("jti")
-            exp = payload.get("exp")
-
-            if not jti or not exp:
-                return False
-
-            # Add to blacklist table
-            with self.audit_manager.transaction() as conn:
-                cursor = conn.cursor()
-
-                # Create blacklist table if not exists (ensure consistent schema)
-                cursor.execute(
-                    """
-                    CREATE TABLE IF NOT EXISTS token_blacklist (
-                        id INTEGER PRIMARY KEY AUTOINCREMENT,
-                        token_hash TEXT NOT NULL UNIQUE,
-                        expires_at TIMESTAMP NOT NULL,
-                        reason TEXT,
-                        created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
-                    )
-                """
-                )
-
-                # Hash token for storage (don't store actual token)
-                token_hash = hashlib.sha256(token.encode()).hexdigest()
-
-                cursor.execute(
-                    """
-                    INSERT OR IGNORE INTO token_blacklist
-                    (token_hash, expires_at, reason)
-                    VALUES (?, ?, ?)
-                """,
-                    (token_hash, datetime.fromtimestamp(exp), reason),
-                )
-
-                conn.commit()
-
-                logger.info(f"Token revoked: {jti[:8]}... (reason: {reason})")
-                return cursor.rowcount > 0
+            token = jwt.encode(claims, self.secret_key, algorithm=self.algorithm)
+            logger.debug(f"User token created for: {username}")
+            return token
 
         except Exception as e:
-            logger.error(f"Failed to revoke token: {e}")
-            return False
-
-    def _create_refresh_token(self, api_key_id: int) -> str:
-        """Create and store refresh token"""
-        try:
-            # Generate secure refresh token
-            refresh_token = secrets.token_urlsafe(64)
-            token_hash = hashlib.sha256(refresh_token.encode()).hexdigest()
-
-            expires_at = datetime.utcnow() + timedelta(
-                days=self.refresh_token_expire_days
-            )
-
-            # Store in database
-            with self.audit_manager.transaction() as conn:
-                cursor = conn.cursor()
-
-                # Ensure the api_key_id exists in api_credentials table
-                # For testing scenarios, create a minimal record if it doesn't exist
-                cursor.execute(
-                    "SELECT id FROM api_credentials WHERE id = ?", (api_key_id,)
-                )
-
-                if not cursor.fetchone():
-                    # Create minimal api_credentials record for testing
-                    cursor.execute(
-                        """
-                        INSERT OR IGNORE INTO api_credentials
-                        (id, service_name, api_key_hash, is_active, created_at, updated_at)
-                        VALUES (?, ?, ?, ?, ?, ?)
-                        """,
-                        (
-                            api_key_id,
-                            f"test_key_{api_key_id}",
-                            "test_hash",
-                            True,
-                            datetime.utcnow(),
-                            datetime.utcnow(),
-                        ),
-                    )
-
-                cursor.execute(
-                    """
-                    INSERT INTO refresh_tokens
-                    (api_key_id, token_hash, expires_at)
-                    VALUES (?, ?, ?)
-                """,
-                    (api_key_id, token_hash, expires_at),
-                )
-
-                conn.commit()
-
-            return refresh_token
-
-        except Exception as e:
-            logger.error(f"Failed to create refresh token: {e}")
+            logger.error(f"Failed to create user token: {e}")
             raise
 
-    def _verify_refresh_token(self, refresh_token: str) -> Optional[int]:
-        """Verify refresh token and return API key ID"""
-        try:
-            token_hash = hashlib.sha256(refresh_token.encode()).hexdigest()
+    def verify_user_token(self, token: str) -> Optional[dict]:
+        """Verify user token and return user info
 
-            with self.audit_manager.transaction() as conn:
-                cursor = conn.cursor()
-
-                cursor.execute(
-                    """
-                    SELECT api_key_id FROM refresh_tokens
-                    WHERE token_hash = ?
-                    AND expires_at > ?
-                    AND is_revoked = 0
-                """,
-                    (token_hash, datetime.utcnow()),
-                )
-
-                row = cursor.fetchone()
-                return row[0] if row else None
-
-        except Exception as e:
-            logger.error(f"Refresh token verification failed: {e}")
-            return None
-
-    def _is_token_blacklisted(self, token: str) -> bool:
-        """Check if token is blacklisted"""
-        try:
-            token_hash = hashlib.sha256(token.encode()).hexdigest()
-
-            with self.audit_manager.transaction() as conn:
-                cursor = conn.cursor()
-
-                cursor.execute(
-                    """
-                    SELECT 1 FROM token_blacklist
-                    WHERE token_hash = ? AND expires_at > ?
-                """,
-                    (token_hash, datetime.utcnow()),
-                )
-
-                return cursor.fetchone() is not None
-
-        except Exception as e:
-            logger.debug(f"Blacklist check failed: {e}")
-            return False
-
-    def cleanup_expired_tokens(self) -> int:
-        """Clean up expired refresh tokens and blacklisted tokens
+        Args:
+            token: JWT token string
 
         Returns:
-            Number of tokens cleaned up
+            Dict with user info if valid, None if invalid
         """
         try:
-            cleaned = 0
-            now = datetime.utcnow()
-
-            with self.audit_manager.transaction() as conn:
-                cursor = conn.cursor()
-
-                # Clean expired refresh tokens
-                cursor.execute(
-                    """
-                    DELETE FROM refresh_tokens WHERE expires_at <= ?
-                """,
-                    (now,),
-                )
-                cleaned += cursor.rowcount
-
-                # Clean expired blacklisted tokens
-                cursor.execute(
-                    """
-                    DELETE FROM token_blacklist WHERE expires_at <= ?
-                """,
-                    (now,),
-                )
-                cleaned += cursor.rowcount
-
-                conn.commit()
-
-            if cleaned > 0:
-                logger.info(f"Cleaned up {cleaned} expired tokens")
-
-            return cleaned
-
+            payload = jwt.decode(token, self.secret_key, algorithms=[self.algorithm])
+            return {
+                "user_id": payload.get("sub"),
+                "username": payload.get("username"),
+                "scopes": payload.get("scopes", ["read"]),
+                "exp": payload.get("exp"),
+            }
+        except (jwt.ExpiredSignatureError, jwt.InvalidTokenError):
+            return None
         except Exception as e:
-            logger.error(f"Token cleanup failed: {e}")
-            return 0
+            logger.error(f"User token verification failed: {e}")
+            return None
+
+    def get_token_info(self, token: str) -> Optional[dict]:
+        """Get basic token information without full verification
+
+        Args:
+            token: JWT token string
+
+        Returns:
+            Dict with token info if decodable, None otherwise
+        """
+        try:
+            # Decode without verification to get info
+            payload = jwt.decode(token, options={"verify_signature": False})
+            return {
+                "subject": payload.get("sub"),
+                "expires_at": datetime.fromtimestamp(payload.get("exp", 0)),
+                "api_key_name": payload.get("api_key_name"),
+                "username": payload.get("username"),
+            }
+        except Exception:
+            return None
+
+    def revoke_token(self, token: str, reason: Optional[str] = None) -> bool:
+        """MVP: Token revocation not implemented (no blacklist for simplicity)
+
+        Returns True for compatibility but doesn't actually revoke anything
+        """
+        logger.info("Token revocation called but not implemented in MVP - Issue #153")
+        return True
+
+    def cleanup_expired_tokens(self) -> int:
+        """MVP: Token cleanup not implemented (no persistent storage)
+
+        Returns 0 for compatibility
+        """
+        logger.info("Token cleanup called but not implemented in MVP - Issue #153")
+        return 0
+
+
+# Convenience functions for backward compatibility
+def create_jwt_manager(
+    db_path: Optional[str] = None, secret_key: Optional[str] = None
+) -> JWTManager:
+    """Create JWTManager instance"""
+    return JWTManager(db_path, secret_key)
+
+
+def get_default_jwt_manager() -> JWTManager:
+    """Get default JWTManager instance"""
+    return JWTManager()
