+++ tests/performance/test_scalability.py
@@ -12,16 +12,57 @@ import psutil
 import pytest
 
 from src.api.production_istat_client import ProductionIstatClient
-from src.services.service_factory import get_dataflow_analysis_service
+
+# get_dataflow_analysis_service removed in Issue #153 (MVP simplification)
 
 
+# Mock objects for removed functionality (tests are skipped anyway)
+class MockService:
+    def __init__(self):
+        self.istat_client = MockClient()
+
+    async def analyze_dataflows_from_xml(self, xml):
+        return MockResult()
+
+    def test_popular_datasets(self):
+        return 0
+
+    def _calculate_priority(self, dataset):
+        return 1.0
+
+    def generate_summary_report(self, data):
+        return "Mock report"
+
+    def _categorize_dataflows_sync(self, data):
+        return {}
+
+
+class MockClient:
+    def __init__(self):
+        self.session = None
+
+
+class MockResult:
+    def __init__(self):
+        self.total_analyzed = 1
+        self.categorized_dataflows = {}
+
+
+# Mock instances for linting (tests are skipped)
+service = MockService()
+adapter = MockService()
+analyzer = MockService()
+
+
+@pytest.mark.skip(
+    reason="Issue #153: get_dataflow_analysis_service removed for MVP - tests disabled temporarily"
+)
 @pytest.mark.performance
 class TestScalabilityPerformance:
     """Test system scalability and performance."""
 
     def test_dataflow_parsing_performance(self, temp_dir):
         """Test dataflow parsing performance with large XML files."""
-        analyzer = get_dataflow_analysis_service()
 
         # Generate large XML file with many dataflows
         large_xml = self._generate_large_dataflow_xml(num_dataflows=1000)
@@ -112,7 +153,7 @@ class TestScalabilityPerformance:
 
     def test_memory_usage_scaling(self, temp_dir):
         """Test memory usage with increasing data sizes."""
-        get_dataflow_analysis_service()
+        # get_dataflow_analysis_service() # Issue #153: removed for MVP
 
         # Test with different data sizes
         data_sizes = [100, 500, 1000, 2000]
@@ -248,7 +289,7 @@ class TestScalabilityPerformance:
 
     def test_categorization_performance(self):
         """Test categorization performance with many datasets."""
-        get_dataflow_analysis_service()
+        # get_dataflow_analysis_service() # Issue #153: removed for MVP
 
         # Generate many datasets
         num_datasets = 5000
@@ -291,7 +332,7 @@ class TestScalabilityPerformance:
 
     def test_batch_processing_performance(self, temp_dir):
         """Test batch processing performance."""
-        get_dataflow_analysis_service()
+        # get_dataflow_analysis_service() # Issue #153: removed for MVP
 
         # Generate batch of datasets
         batch_size = 100
