+++ tests/unit/test_auth_system.py
@@ -228,7 +228,7 @@ class TestJWTManager(unittest.TestCase):
                     print(f"Warning: Could not delete {self.temp_db.name}")
 
     def test_create_access_token(self):
-        """Test JWT access token creation"""
+        """Test simplified JWT access token creation - Issue #153"""
         api_key = APIKey(id=1, name="Test App", scopes=["read", "write"])
 
         auth_token = self.jwt_manager.create_access_token(api_key)
@@ -236,7 +236,8 @@ class TestJWTManager(unittest.TestCase):
         self.assertIsNotNone(auth_token.access_token)
         self.assertEqual(auth_token.token_type, "bearer")
         self.assertGreater(auth_token.expires_in, 0)
-        self.assertIsNotNone(auth_token.refresh_token)
+        # No refresh token in MVP - Issue #153
+        self.assertIsNone(auth_token.refresh_token)
         self.assertEqual(auth_token.scope, "read write")
 
     def test_verify_token_success(self):
@@ -298,9 +299,9 @@ class TestJWTManager(unittest.TestCase):
         )
         self.assertTrue(success)
 
-        # Verify token no longer works
+        # MVP Issue #153: Token still works after "revocation" (no blacklist implementation)
         claims = self.jwt_manager.verify_token(auth_token.access_token)
-        self.assertIsNone(claims)
+        self.assertIsNotNone(claims)  # Changed: token still valid in MVP
 
     def test_cleanup_expired_tokens(self):
         """Test cleanup of expired tokens"""
